--- app/uk/gov/hmrc/pillar2/connectors/headers/UKTaxReturnHeaders.scala
@@ -1,36 +1,0 @@
-/*
- * Copyright 2024 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package uk.gov.hmrc.pillar2.connectors.headers
-
-import java.time.ZonedDateTime
-import java.time.format.DateTimeFormatter
-import java.util.UUID
-
-trait UKTaxReturnHeaders {
-  def generateCorrelationId(): String = UUID.randomUUID().toString
-
-  def generateReceiptDate(): String =
-    ZonedDateTime.now().format(DateTimeFormatter.ISO_INSTANT)
-
-  def generateHeaders(pillar2Id: String): Seq[(String, String)] = Seq(
-    "correlationid"         -> generateCorrelationId(),
-    "X-Originating-System"  -> "MDTP",
-    "X-Pillar2-Id"          -> pillar2Id,
-    "X-Receipt-Date"        -> generateReceiptDate(),
-    "X-Transmitting-System" -> "HIP"
-  )
-}
