+++ app/uk/gov/hmrc/pillar2/controllers/package.scala
@@ -22,6 +22,8 @@ import play.api.libs.json.{JsError, JsSuccess, Json}
 import play.api.mvc.Result
 import play.api.mvc.Results._
 import uk.gov.hmrc.http.HttpResponse
+import uk.gov.hmrc.pillar2.models.hip.ErrorSummary.result_500
+import uk.gov.hmrc.pillar2.models.hip.{ErrorSummary, SuccessResponse, ValidationErrors}
 import uk.gov.hmrc.pillar2.models.hods.ErrorDetails
 
 import scala.util.{Success, Try}
@@ -76,6 +78,22 @@ package object controllers {
         logDownStreamError(httpResponse.body)
         InternalServerError(httpResponse.body)
 
+    }
+
+  private[controllers] def convertToApiResult(response: HttpResponse): Result =
+    response.status match {
+      case 201 =>
+        response.json.validate[SuccessResponse] match {
+          case JsSuccess(success, _) => Ok(Json.toJson(success))
+          case JsError(_)            => result_500
+        }
+      case 422 =>
+        response.json.validate[ValidationErrors] match {
+          case JsSuccess(validationErrors, _) => UnprocessableEntity(Json.toJson(ErrorSummary(validationErrors.code, validationErrors.text)))
+          case JsError(_)                     => result_500
+        }
+      case _ =>
+        result_500 // TODO: This loses a lot of infomation on what the error actually is and we rely on the implicit logs provided by play logging, maybe set this up to decode the message
     }
 
   private def logDownStreamError(
